+++ tests/HotelRepositoryTest.php
@@ -6,13 +6,16 @@ use Dotenv\Dotenv;
 use PHPUnit\Framework\TestCase;
 use Towa\SDK\Bookingcom\Repository\Hotel_Repository;
 use Towa\SDK\Bookingcom\Repository\City_Repository;
+use Towa\SDK\Bookingcom\Repository\Facility_Repository;
 use Towa\SDK\Bookingcom\Model\Hotel;
 use Towa\SDK\Bookingcom\Model\Hotel_Type;
 
 class HotelRepositoryTest extends TestCase
 {
     /** @var \Towa\SDK\Bookingcom\Repository\Hotel_Repository */
+    /** @var \Towa\SDK\Bookingcom\Repository\Facility_Repository */
     protected $hotelRepo;
+    protected $facilityRepo;
 
     public function setUp()
     {
@@ -20,6 +23,7 @@ class HotelRepositoryTest extends TestCase
         $dotenv->load();
 
         $this->hotelRepo = new Hotel_Repository($_ENV['USERNAME'], $_ENV['PASSWORD']);
+        $this->facilityRepo = new Facility_Repository($_ENV['USERNAME'], $_ENV['PASSWORD']);
 
         parent::setUp();
     }
@@ -35,6 +39,7 @@ class HotelRepositoryTest extends TestCase
         );
 
         $this->assertNotEmpty($hotels);
+        $this->assertCount(1, [$hotels]);
     }
 
     /** @test */
@@ -43,7 +48,7 @@ class HotelRepositoryTest extends TestCase
         $hotels = $this->hotelRepo->get_hotels(
             [
                 'hotel_ids' => 28546,
-                'extras' => 'hotel_info',
+                'extras' => 'hotel_info,hotel_description,hotel_facilities,hotel_photos,room_info',
             ],
             'en'
         );
@@ -98,5 +103,44 @@ class HotelRepositoryTest extends TestCase
         $this->assertNotEmpty($hotel);
         $this->assertCount(1, [$hotels]);
         // $this->assertInstanceOf(Hotel::class, $hotels);
+    }
+
+
+    /** @test */
+    public function it_can_get_changed_hotels()
+    {
+        $hotels = $this->hotelRepo->get_changed_hotels(
+                 [
+                    'last_change'=>'2018-07-09 00:00:00',
+                ]
+             );
+           
+        $hotel = $hotels[0]->hotel_changed_id();
+        $this->assertNotEmpty($hotel);
+        $this->assertCount(1, [$hotels]);
+        // $this->assertInstanceOf(Hotel::class, $hotels);
+    }
+
+    /** @test */
+    public function it_can_get_closed_hotels()
+    {
+        $hotels = $this->hotelRepo->get_changed_hotels(
+                  [
+                     'last_change'=>'2018-07-09 00:00:00',
+                 ]
+              );
+            
+        $hotel = $hotels[0]->hotel_closed_id();
+        $this->assertNotEmpty($hotel);
+        $this->assertCount(1, [$hotels]);
+    }
+    
+    /** @test */
+    public function it_can_get_facility_types()
+    {
+        $facilities = $this->facilityRepo->get_facility_types([]);
+            
+        $facility = $facilities;
+        $this->assertNotEmpty($facility);
     }
 }